Bugfix render template position (#60)
* use go:embed to load tmpl files from binary

* fix bug remove already exist dir

+++ internal/usecase/generator/off/generate_offline.go
@@ -2,6 +2,7 @@ package off
 
 import (
 	"fmt"
+	"io/fs"
 	"os"
 	"path/filepath"
 	"regexp"
@@ -23,9 +24,10 @@ func (u *usecase) Generate(inputEntity generator.GeneratorInputEntity) error {
 	}
 
 	var err error
+	var fileInfo fs.FileInfo
 	defer func() {
-		// rollback if got any error
-		if err != nil {
+		// rollback if got any error except getting error of root directory already exist
+		if err != nil && !fileInfo.IsDir() {
 			if err := os.RemoveAll(inputEntity.ProjectName); err != nil {
 				rolling.New().Error("rollback failed", zap.Error(err))
 			}
@@ -33,7 +35,7 @@ func (u *usecase) Generate(inputEntity generator.GeneratorInputEntity) error {
 	}()
 
 	// create root directory for the project
-	if err = u.createDir(inputEntity.ProjectName); err != nil {
+	if fileInfo, err = u.createDir(inputEntity.ProjectName); err != nil {
 		return fmt.Errorf("failed to create project directory: %s", err.Error())
 	}
 
@@ -48,7 +50,7 @@ func (u *usecase) createChildDirectories(inputEntity generator.GeneratorInputEnt
 			if err := u.repo.RenderTemplate(generator.GeneratorInputEntity{
 				ProjectName:    inputEntity.ProjectName,
 				ModuleName:     inputEntity.ModuleName,
-				TempFilePath:   filepath.Join("templates", path, key),
+				TempFilePath:   filepath.Join(generator.OFF_TEMP_DIR_NAME, path, key),
 				TargetFilePath: filepath.Join(inputEntity.ProjectName, absFileName),
 			}); err != nil {
 				return err
@@ -65,14 +67,14 @@ func (u *usecase) createChildDirectories(inputEntity generator.GeneratorInputEnt
 	return nil
 }
 
-func (u *usecase) createDir(dirName string) error {
+func (u *usecase) createDir(dirName string) (fs.FileInfo, error) {
 	info, err := os.Stat(dirName)
 	if os.IsNotExist(err) {
-		return os.Mkdir(dirName, 0755)
+		return info, os.Mkdir(dirName, 0755)
 	} else if err == nil && info.IsDir() {
-		return fmt.Errorf("directory already exists: %s", dirName)
+		return info, fmt.Errorf("directory already exists: %s", dirName)
 	} else {
-		return fmt.Errorf("error checking directory: %s", err.Error())
+		return info, fmt.Errorf("error checking directory: %s", err.Error())
 	}
 }
